requirements.txt.txt
streamlit
tensorflow
numpy
opencv-python-headless
Pillow